+++ CTC/Home/Calendar/Model/Event.swift
@@ -1,0 +1,17 @@
+//
+//  Event.swift
+//  CTC
+//
+//  Created by Jared Williams on 2/4/19.
+//  Copyright © 2019 Code That Cares. All rights reserved.
+//
+
+import Foundation
+
+struct Event {
+    var time: String?
+    var location: String?
+    var length: Int?
+    var title: String?
+    var date: Date?
+}
