+++ CTC/Home/Calendar/View/EventCellViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  EventCellViewModel.swift
+//  CTC
+//
+//  Created by Jared Williams on 2/4/19.
+//  Copyright © 2019 Code That Cares. All rights reserved.
+//
+
+import Foundation
+
+class EventCellViewModel {
+    
+    private var event: Event?
+    init(event: Event) {
+        
+    }
+}
